[FIX]: Add empty new lines.
Added missing new empty lines to the file.
Because before this PR there is mix of no empty new lines and new empty lines.
And it won't render properly on a terminal, and it's considered a bad practice.
Add a simple line feed as the last character to fix it.

Fix: #16

Author: Tim Joosten <Topairy@gmail.com>